Some minor post-course adjustments
Gave the product service the ability to return one
product by id. Required a little digging into
Observable in RxJs to understand how to hook that up.
Confusing at first but very simple in the end.
Also hooked the star component into the detail page
and tweaked a few things there, too.
This is probably done now... will move on to
additional tutorials or maybe start a project.